requirements.txt
langchain==0.1.0
langgraph==0.0.20
langsmith==0.0.69
streamlit==1.28.1
finnhub-python==2.4.18
boto3==1.34.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
asyncio-mqtt==0.16.1
aiohttp==3.9.1
plotly==5.17.0
yfinance==0.2.28
scikit-learn==1.3.2
ta==0.10.2
